update to the original repo (#2)
* Change circleci mongo version to one that still exists

* brewRenderer: Capture event properties while still in event loop.

React [pools events](https://github.com/facebook/react/issues/2850),
which means that once the lambda runs later on the properties we want
to read will be null. Instead, we capture the properties in the event
loop.

* Add dockerization and example compose file

* Update according to change request

* Add missing volume declaration.

* s/let/const/ to satisfy ESLint

* Fix border-image bug on PDFs

I think this might be caused by a new bug in Chrome or something since it seems to have popped up in GMBinder as well in the last couple weeks. See #683

Anyway, the error only occurs on the Class Table and Descriptive Text Box, which are also the only ones to use `border-image-repeat: round`. I changed it to `border-image-repeat: stretch` which solves the graphical issue and doesn't affect the appearance otherwise in this particular case.

* Update README.md

Typos and additional license to sync up with http://homebrewery.naturalcrit.com.

* Fixed typo "is" -> "in"

* Fixing issue#413 class feature duping entire brew

* Re-enable Box Shadows in PDF

Box shadows have been fixed in Chrome. No reason to hide them anymore.

* Adding circleCI

* Adding circleCI and eslint

* adding a circleCI task

* Adding in test case runner

* Moving babelrc into package

* Updating contributing and issue guides, added new verify npm task

* Adding mongo image to the circleci config

* Updating navbar link to go to the new subreddit

* lint

* Updating build scripts and libraries

* Updating build scripts and libraries

* Adding in the missing repository field

* Updating superagent

* Upping the node ver on circleci

* Updating to user create-react-class

* Upping version

* Updating some packages

* adding bithoundrc and updating pico-router

* Removing picofluc for now

* upgrading to mongoose v5

* Change circleci mongo version to one that still exists

* Fix regression with updated Pico-Router

This variable name changed with the v2.0.0 of Pico-Router and it wasn't changed here. This was causing every page to display a  `<div class="homePage page">` rather than the appropriate `editPage` or `sharePage`. It just wasn't immediately noticeable since `homePage` and `editPage` are almost identical. The problem is clearly visible on `sharePage` however.

`homebrew.jsx` was set to always render a defaultUrl of ` ` which creates a `<div class="homePage page">` by default, overwriting the expected `<div>` on for the current url.

* Fix metadata not working. Some earlier linting caused the handleFieldChange and handleSystem functions to not have access to event data. Returning these to their pre-lint state for now, but there might be a more lint-friendly way to do this later.

* Found the lint-friendly way to do it. Just needed to pass along the event data.

* Added spaces for lint...

* Fix metadata not working. Some earlier linting caused the handleFieldChange and handleSystem functions to not have access to event data. Returning these to their pre-lint state for now, but there might be a more lint-friendly way to do this later.

* Found the lint-friendly way to do it. Just needed to pass along the event data.

* Added spaces for lint...

* Wrap the text in the "source" page.

See #332.

* Hide duplicate scrollbar in brewRenderer

* Add bold/italic shortcuts, reposition renderer warning

* Reset to previous commit

* Fix ESLint errors

* Make imgur link https in order to display image

Newer browsers warn and/or do not display content that gets loaded via HTTP if site is visited via HTTPS.

* Adding some more text to the default issue title to guide peopel to add an issue title

* Added in a function to sanatize script tags specifically

* Making the script tag stripper case insensitive

* Added middleware to force all routes to be https

* Updating readme

* Finished stats and brew lookup on admin panel

* Removing the admin key, switching to full basic auth

* Finished the clean up brew ui

* lint

* Added a notification popup similar to the "YOU ARE NOT USING CHROME" message. Can be used to notify users of current known issues, updates, etc.

Currently just reminds users to back up documents with a link to the Reddit FAQ.

* Added a notification popup similar to the "YOU ARE NOT USING CHROME" message. Can be used to notify users of current known issues, updates, etc.

Currently just reminds users to back up documents with a link to the Reddit FAQ.

* Refactored the "Recent Items" navbar component. Greatly simplified the code to about 70% length and made it much easier to read. Results in the same thing.

Oh. And it works again. Recent brews should be showing up fine now.

* Lint

* More Lint

* Fix duplicate scrollbar in "share" page

* Revert "Merge branch 'master' of https://github.com/naturalcrit/homebrewery"

This reverts commit 077511dfa7fdfbc3d2b59bac2cb9949da003d16c, reversing
changes made to facbc5f6dcb97067b8d25d21bfd7eb1d44c78bda.

* Revert "Revert "Merge branch 'master' of https://github.com/naturalcrit/homebrewery""

This reverts commit 0c70162a786cfab26f39b43ab12dd363ece1c6a4.

* Hide scrollbar in share page but KEEP SCROLLBAR IN USER PAGE

* Fix typo in changelog

Fixes #765

* More version updates in the changelog

* Make Monster Blocks always generate "abilities"

Originally had a random chance to generate no abilities, which a user might not know how to insert with the proper formatting. Also added some new abilities to select randomly when generated. Also fixes #836 .

* Notification about Chrome PDF missing letters bug

Changed DISMISS KEY to current date to make popup appear again.

* Fix spacing on Notification popup for legibility

* Fixing delete buttons

Delete should now first remove the current user as an editor, and only delete the file if he was the last editor. Currently, anyone who views your brew is added as an editor and if they delete it, the whole brew is deleted for everyone.

* Lint

* Change popup warning if you are the only editor

If you are the only editor, it will warn you that the document is about to be permanently deleted.

* Change the popup warning on the metadata editor

* Lint

* Lint 2